more (more) `uint*()`

+++ chia/simulator/block_tools.py
@@ -138,28 +138,28 @@ DESERIALIZE_MOD = load_serialized_clvm_maybe_recompile(
 test_constants = dataclasses.replace(
     DEFAULT_CONSTANTS,
     MIN_PLOT_SIZE=18,
-    MIN_BLOCKS_PER_CHALLENGE_BLOCK=12,
-    DIFFICULTY_STARTING=2**10,
+    MIN_BLOCKS_PER_CHALLENGE_BLOCK=uint8(12),
+    DIFFICULTY_STARTING=uint64(2**10),
     DISCRIMINANT_SIZE_BITS=16,
-    SUB_EPOCH_BLOCKS=170,
-    WEIGHT_PROOF_THRESHOLD=2,
-    WEIGHT_PROOF_RECENT_BLOCKS=380,
-    DIFFICULTY_CONSTANT_FACTOR=33554432,
-    NUM_SPS_SUB_SLOT=16,  # Must be a power of 2
-    MAX_SUB_SLOT_BLOCKS=50,
-    EPOCH_BLOCKS=340,
-    BLOCKS_CACHE_SIZE=340 + 3 * 50,  # Coordinate with the above values
+    SUB_EPOCH_BLOCKS=uint32(170),
+    WEIGHT_PROOF_THRESHOLD=uint8(2),
+    WEIGHT_PROOF_RECENT_BLOCKS=uint32(380),
+    DIFFICULTY_CONSTANT_FACTOR=uint128(33554432),
+    NUM_SPS_SUB_SLOT=uint32(16),  # Must be a power of 2
+    MAX_SUB_SLOT_BLOCKS=uint32(50),
+    EPOCH_BLOCKS=uint32(340),
+    BLOCKS_CACHE_SIZE=uint32(340 + 3 * 50),  # Coordinate with the above values
     SUB_SLOT_TIME_TARGET=600,  # The target number of seconds per slot, mainnet 600
-    SUB_SLOT_ITERS_STARTING=2**10,  # Must be a multiple of 64
+    SUB_SLOT_ITERS_STARTING=uint64(2**10),  # Must be a multiple of 64
     NUMBER_ZERO_BITS_PLOT_FILTER=1,  # H(plot signature of the challenge) must start with these many zeroes
     # Allows creating blockchains with timestamps up to 10 days in the future, for testing
     MAX_FUTURE_TIME2=3600 * 24 * 10,
     MEMPOOL_BLOCK_BUFFER=6,
-    UNIQUE_PLOTS_WINDOW=2,
+    UNIQUE_PLOTS_WINDOW=uint8(2),
     # we deliberately make this different from HARD_FORK_HEIGHT in the
     # tests, to ensure they operate independently (which they need to do for
     # testnet10)
-    HARD_FORK_FIX_HEIGHT=5496100,
+    HARD_FORK_FIX_HEIGHT=uint32(5496100),
 )
